Refactor also position constructor method
