+++ src/Lynx.Benchmark/MakeUnmakeMove_implementation_Benchmark.cs
@@ -1,5 +1,4 @@
 ﻿/*
- * Consistent local reults
  *
  *  BenchmarkDotNet v0.13.12, Ubuntu 22.04.3 LTS (Jammy Jellyfish)
  *  AMD EPYC 7763, 1 CPU, 4 logical and 2 physical cores
@@ -7,44 +6,43 @@
  *    [Host]     : .NET 8.0.1 (8.0.123.58001), X64 RyuJIT AVX2
  *    DefaultJob : .NET 8.0.1 (8.0.123.58001), X64 RyuJIT AVX2
  *
- *
- *  | Method                                             | data                 | Mean         | Error       | StdDev      | Ratio | RatioSD | Gen0       | Allocated     | Alloc Ratio |
- *  |--------------------------------------------------- |--------------------- |-------------:|------------:|------------:|------:|--------:|-----------:|--------------:|------------:|
- *  | NewPosition                                        | (2K2r(...)1, 6) [38] | 371,916.2 us | 3,357.88 us | 2,976.67 us |  1.00 |    0.00 | 13000.0000 | 1069286.11 KB |        1.00 |
- *  | MakeUnmakeMove_Original                            | (2K2r(...)1, 6) [38] | 438,803.6 us | 3,102.83 us | 2,902.39 us |  1.18 |    0.01 |  3000.0000 |  278498.13 KB |        0.26 |
- *  | MakeUnmakeMove_WithZobristKey                      | (2K2r(...)1, 6) [38] | 276,965.8 us | 2,425.10 us | 2,268.44 us |  0.75 |    0.01 |  3000.0000 |  278497.61 KB |        0.26 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (2K2r(...)1, 6) [38] |  21,992.5 us |   181.89 us |   161.24 us |  0.06 |    0.00 |   406.2500 |   35056.14 KB |        0.03 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (2K2r(...)1, 6) [38] |  22,462.7 us |   394.02 us |   349.29 us |  0.06 |    0.00 |   406.2500 |   35056.14 KB |        0.03 |
- *  |                                                    |                      |              |             |             |       |         |            |               |             |
- *  | NewPosition                                        | (3K4/(...)1, 6) [38] | 388,579.9 us | 3,590.55 us | 3,358.60 us |  1.00 |    0.00 | 13000.0000 | 1069286.11 KB |       1.000 |
- *  | MakeUnmakeMove_Original                            | (3K4/(...)1, 6) [38] | 441,682.8 us | 3,460.67 us | 3,067.79 us |  1.14 |    0.02 |  3000.0000 |  278498.13 KB |       0.260 |
- *  | MakeUnmakeMove_WithZobristKey                      | (3K4/(...)1, 6) [38] | 260,931.1 us | 2,110.65 us | 1,974.30 us |  0.67 |    0.01 |  3000.0000 |  278497.61 KB |       0.260 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (3K4/(...)1, 6) [38] |  21,348.5 us |   129.00 us |   120.66 us |  0.05 |    0.00 |    62.5000 |    5446.33 KB |       0.005 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (3K4/(...)1, 6) [38] |  21,358.5 us |   151.00 us |   141.25 us |  0.05 |    0.00 |    62.5000 |    5446.33 KB |       0.005 |
- *  |                                                    |                      |              |             |             |       |         |            |               |             |
- *  | NewPosition                                        | (8/p7(...)-, 6) [37] |  13,754.5 us |    54.57 us |    51.05 us |  1.00 |    0.00 |   515.6250 |    42915.8 KB |        1.00 |
- *  | MakeUnmakeMove_Original                            | (8/p7(...)-, 6) [37] |  15,711.2 us |    26.07 us |    24.39 us |  1.14 |    0.00 |   187.5000 |   17113.08 KB |        0.40 |
- *  | MakeUnmakeMove_WithZobristKey                      | (8/p7(...)-, 6) [37] |  10,406.1 us |    55.41 us |    49.12 us |  0.76 |    0.01 |   203.1250 |   17113.07 KB |        0.40 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (8/p7(...)-, 6) [37] |   4,618.6 us |    25.61 us |    23.96 us |  0.34 |    0.00 |   101.5625 |    8398.54 KB |        0.20 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (8/p7(...)-, 6) [37] |   4,712.0 us |    52.73 us |    49.32 us |  0.34 |    0.00 |   101.5625 |    8398.54 KB |        0.20 |
- *  |                                                    |                      |              |             |             |       |         |            |               |             |
- *  | NewPosition                                        | (r3k2(...)1, 4) [73] | 319,595.5 us | 1,843.29 us | 1,724.21 us | 1.000 |    0.00 |  9500.0000 |  801233.01 KB |       1.000 |
- *  | MakeUnmakeMove_Original                            | (r3k2(...)1, 4) [73] | 347,561.9 us | 1,235.28 us |   964.43 us | 1.087 |    0.01 |  1000.0000 |  104639.41 KB |       0.131 |
- *  | MakeUnmakeMove_WithZobristKey                      | (r3k2(...)1, 4) [73] | 244,314.2 us |   924.89 us |   819.89 us | 0.764 |    0.00 |  1000.0000 |  104638.72 KB |       0.131 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (r3k2(...)1, 4) [73] |     161.3 us |     0.43 us |     0.33 us | 0.001 |    0.00 |     1.2207 |     102.76 KB |       0.000 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (r3k2(...)1, 4) [73] |     168.9 us |     0.87 us |     0.77 us | 0.001 |    0.00 |     1.2207 |     102.76 KB |       0.000 |
- *  |                                                    |                      |              |             |             |       |         |            |               |             |
- *  | NewPosition                                        | (r4rk(...)0, 4) [77] | 307,631.2 us | 2,571.62 us | 2,405.49 us | 1.000 |    0.00 |  9000.0000 |  765557.21 KB |       1.000 |
- *  | MakeUnmakeMove_Original                            | (r4rk(...)0, 4) [77] | 279,892.9 us | 1,750.95 us | 1,637.84 us | 0.910 |    0.01 |  1000.0000 |   96332.98 KB |       0.126 |
- *  | MakeUnmakeMove_WithZobristKey                      | (r4rk(...)0, 4) [77] | 215,268.1 us | 1,380.54 us | 1,291.35 us | 0.700 |    0.01 |  1000.0000 |   96332.81 KB |       0.126 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (r4rk(...)0, 4) [77] |     347.8 us |     1.24 us |     1.10 us | 0.001 |    0.00 |     2.4414 |      208.5 KB |       0.000 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (r4rk(...)0, 4) [77] |     358.4 us |     1.61 us |     1.26 us | 0.001 |    0.00 |     2.4414 |      208.5 KB |       0.000 |
- *  |                                                    |                      |              |             |             |       |         |            |               |             |
- *  | NewPosition                                        | (rnbq(...)1, 4) [61] |  16,786.2 us |    34.14 us |    28.51 us |  1.00 |    0.00 |   531.2500 |   43733.27 KB |       1.000 |
- *  | MakeUnmakeMove_Original                            | (rnbq(...)1, 4) [61] |  15,329.7 us |    24.72 us |    20.64 us |  0.91 |    0.00 |   109.3750 |    9760.58 KB |       0.223 |
- *  | MakeUnmakeMove_WithZobristKey                      | (rnbq(...)1, 4) [61] |  12,382.0 us |    41.57 us |    36.85 us |  0.74 |    0.00 |   109.3750 |    9760.58 KB |       0.223 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (rnbq(...)1, 4) [61] |     250.9 us |     1.71 us |     1.51 us |  0.01 |    0.00 |     2.4414 |     220.02 KB |       0.005 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (rnbq(...)1, 4) [61] |     249.2 us |     1.73 us |     1.62 us |  0.01 |    0.00 |     2.4414 |     220.02 KB |       0.005 |
+ *  | Method                                             | data                 | Mean         | Error       | StdDev      | Ratio | Gen0       | Allocated     | Alloc Ratio |
+ *  |--------------------------------------------------- |--------------------- |-------------:|------------:|------------:|------:|-----------:|--------------:|------------:|
+ *  | NewPosition                                        | (2K2r(...)1, 6) [38] | 397,636.8 us | 1,175.57 us |   917.81 us |  1.00 | 13000.0000 | 1069286.11 KB |        1.00 |
+ *  | MakeUnmakeMove_Original                            | (2K2r(...)1, 6) [38] | 440,829.8 us | 2,429.95 us | 2,272.97 us |  1.11 |  3000.0000 |  278498.13 KB |        0.26 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (2K2r(...)1, 6) [38] | 280,883.6 us | 3,071.18 us | 2,722.52 us |  0.71 |  3000.0000 |  278497.61 KB |        0.26 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (2K2r(...)1, 6) [38] |  22,239.5 us |    93.62 us |    82.99 us |  0.06 |   406.2500 |   35056.14 KB |        0.03 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (2K2r(...)1, 6) [38] |  21,462.0 us |   131.14 us |   122.67 us |  0.05 |   406.2500 |   35056.14 KB |        0.03 |
+ *  |                                                    |                      |              |             |             |       |            |               |             |
+ *  | NewPosition                                        | (3K4/(...)1, 6) [38] | 392,544.0 us | 1,401.80 us | 1,094.43 us |  1.00 | 13000.0000 | 1069286.11 KB |       1.000 |
+ *  | MakeUnmakeMove_Original                            | (3K4/(...)1, 6) [38] | 442,075.5 us | 3,344.91 us | 2,965.17 us |  1.13 |  3000.0000 |  278498.13 KB |       0.260 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (3K4/(...)1, 6) [38] | 284,606.2 us | 1,503.41 us | 1,332.74 us |  0.73 |  3000.0000 |  278497.61 KB |       0.260 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (3K4/(...)1, 6) [38] |  21,033.2 us |    32.48 us |    25.36 us |  0.05 |    62.5000 |    5446.33 KB |       0.005 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (3K4/(...)1, 6) [38] |  20,827.3 us |    99.09 us |    92.69 us |  0.05 |    62.5000 |    5446.33 KB |       0.005 |
+ *  |                                                    |                      |              |             |             |       |            |               |             |
+ *  | NewPosition                                        | (8/p7(...)-, 6) [37] |  14,260.2 us |    46.97 us |    43.94 us |  1.00 |   515.6250 |    42915.8 KB |        1.00 |
+ *  | MakeUnmakeMove_Original                            | (8/p7(...)-, 6) [37] |  16,042.8 us |    43.61 us |    34.05 us |  1.12 |   187.5000 |   17113.08 KB |        0.40 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (8/p7(...)-, 6) [37] |  10,309.4 us |    28.15 us |    24.95 us |  0.72 |   203.1250 |   17113.07 KB |        0.40 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (8/p7(...)-, 6) [37] |   4,650.9 us |    29.48 us |    27.58 us |  0.33 |   101.5625 |    8398.54 KB |        0.20 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (8/p7(...)-, 6) [37] |   4,650.4 us |    30.97 us |    27.46 us |  0.33 |   101.5625 |    8398.54 KB |        0.20 |
+ *  |                                                    |                      |              |             |             |       |            |               |             |
+ *  | NewPosition                                        | (r3k2(...)1, 4) [73] | 329,451.7 us | 1,566.86 us | 1,465.64 us | 1.000 |  9500.0000 |  801233.01 KB |       1.000 |
+ *  | MakeUnmakeMove_Original                            | (r3k2(...)1, 4) [73] | 331,610.6 us |   419.38 us |   327.42 us | 1.006 |  1000.0000 |  104639.41 KB |       0.131 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (r3k2(...)1, 4) [73] | 249,777.8 us | 2,035.13 us | 1,903.66 us | 0.758 |  1000.0000 |  104638.89 KB |       0.131 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (r3k2(...)1, 4) [73] |     163.1 us |     0.33 us |     0.27 us | 0.000 |     1.2207 |     102.76 KB |       0.000 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (r3k2(...)1, 4) [73] |     161.1 us |     2.51 us |     2.35 us | 0.000 |     1.2207 |     102.76 KB |       0.000 |
+ *  |                                                    |                      |              |             |             |       |            |               |             |
+ *  | NewPosition                                        | (r4rk(...)0, 4) [77] | 310,900.6 us | 1,324.89 us | 1,034.39 us | 1.000 |  9000.0000 |  765557.21 KB |       1.000 |
+ *  | MakeUnmakeMove_Original                            | (r4rk(...)0, 4) [77] | 282,021.9 us | 1,572.90 us | 1,471.30 us | 0.907 |  1000.0000 |   96332.98 KB |       0.126 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (r4rk(...)0, 4) [77] | 213,212.4 us | 1,601.03 us | 1,497.60 us | 0.685 |  1000.0000 |   96332.81 KB |       0.126 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (r4rk(...)0, 4) [77] |     348.3 us |     2.25 us |     2.11 us | 0.001 |     2.4414 |      208.5 KB |       0.000 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (r4rk(...)0, 4) [77] |     340.1 us |     1.96 us |     1.74 us | 0.001 |     2.4414 |      208.5 KB |       0.000 |
+ *  |                                                    |                      |              |             |             |       |            |               |             |
+ *  | NewPosition                                        | (rnbq(...)1, 4) [61] |  17,400.8 us |   200.22 us |   187.28 us |  1.00 |   531.2500 |   43733.27 KB |       1.000 |
+ *  | MakeUnmakeMove_Original                            | (rnbq(...)1, 4) [61] |  15,707.5 us |    22.23 us |    19.71 us |  0.90 |    93.7500 |    9760.59 KB |       0.223 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (rnbq(...)1, 4) [61] |  11,660.6 us |    53.70 us |    50.23 us |  0.67 |   109.3750 |    9760.58 KB |       0.223 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (rnbq(...)1, 4) [61] |     228.4 us |     0.39 us |     0.36 us |  0.01 |     2.6855 |     220.02 KB |       0.005 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (rnbq(...)1, 4) [61] |     247.4 us |     1.68 us |     1.57 us |  0.01 |     2.4414 |     220.02 KB |       0.005 |
  *
  *
  *  BenchmarkDotNet v0.13.12, Windows 10 (10.0.20348.2159) (Hyper-V)
@@ -55,86 +53,87 @@
  *
  *  | Method                                             | data                 | Mean         | Error       | StdDev      | Ratio | RatioSD | Gen0       | Gen1      | Allocated     | Alloc Ratio |
  *  |--------------------------------------------------- |--------------------- |-------------:|------------:|------------:|------:|--------:|-----------:|----------:|--------------:|------------:|
- *  | NewPosition                                        | (2K2r(...)1, 6) [38] | 340,274.7 us | 5,376.88 us | 5,029.54 us |  1.00 |    0.00 | 65000.0000 | 1000.0000 | 1069285.78 KB |        1.00 |
- *  | MakeUnmakeMove_Original                            | (2K2r(...)1, 6) [38] | 442,876.5 us | 4,237.68 us | 3,538.65 us |  1.30 |    0.01 | 17000.0000 |         - |   278497.8 KB |        0.26 |
- *  | MakeUnmakeMove_WithZobristKey                      | (2K2r(...)1, 6) [38] | 269,098.4 us |   554.73 us |   491.76 us |  0.79 |    0.01 | 17000.0000 |         - |  278497.45 KB |        0.26 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (2K2r(...)1, 6) [38] |  21,627.3 us |   416.92 us |   480.13 us |  0.06 |    0.00 |  2125.0000 |   31.2500 |   35057.84 KB |        0.03 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (2K2r(...)1, 6) [38] |  22,258.5 us |   383.98 us |   359.18 us |  0.07 |    0.00 |  2125.0000 |   31.2500 |   35057.84 KB |        0.03 |
+ *  | NewPosition                                        | (2K2r(...)1, 6) [38] | 362,623.0 us | 7,118.07 us | 8,197.18 us |  1.00 |    0.00 | 65000.0000 | 1000.0000 | 1069285.13 KB |        1.00 |
+ *  | MakeUnmakeMove_Original                            | (2K2r(...)1, 6) [38] | 469,134.5 us | 7,765.80 us | 7,264.13 us |  1.29 |    0.04 | 17000.0000 |         - |   278497.8 KB |        0.26 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (2K2r(...)1, 6) [38] | 262,467.8 us | 5,146.64 us | 7,703.25 us |  0.73 |    0.02 | 17000.0000 |         - |  278497.45 KB |        0.26 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (2K2r(...)1, 6) [38] |  21,355.4 us |   155.29 us |   145.26 us |  0.06 |    0.00 |  2125.0000 |   31.2500 |   35057.84 KB |        0.03 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (2K2r(...)1, 6) [38] |  21,132.3 us |   205.52 us |   192.24 us |  0.06 |    0.00 |  2125.0000 |   31.2500 |   35057.84 KB |        0.03 |
  *  |                                                    |                      |              |             |             |       |         |            |           |               |             |
- *  | NewPosition                                        | (3K4/(...)1, 6) [38] | 345,690.7 us | 6,394.45 us | 5,981.37 us |  1.00 |    0.00 | 65000.0000 | 1000.0000 | 1069285.78 KB |       1.000 |
- *  | MakeUnmakeMove_Original                            | (3K4/(...)1, 6) [38] | 437,070.3 us | 2,823.89 us | 2,358.08 us |  1.26 |    0.02 | 17000.0000 |         - |   278497.8 KB |       0.260 |
- *  | MakeUnmakeMove_WithZobristKey                      | (3K4/(...)1, 6) [38] | 270,378.5 us | 2,361.96 us | 2,209.38 us |  0.78 |    0.01 | 17000.0000 |         - |  278497.45 KB |       0.260 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (3K4/(...)1, 6) [38] |  21,175.1 us |    76.07 us |    71.15 us |  0.06 |    0.00 |   312.5000 |         - |    5447.08 KB |       0.005 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (3K4/(...)1, 6) [38] |  21,442.6 us |    72.25 us |    64.05 us |  0.06 |    0.00 |   312.5000 |         - |    5447.08 KB |       0.005 |
+ *  | NewPosition                                        | (3K4/(...)1, 6) [38] | 317,801.2 us | 3,109.67 us | 2,908.79 us |  1.00 |    0.00 | 65000.0000 | 1000.0000 | 1069285.42 KB |       1.000 |
+ *  | MakeUnmakeMove_Original                            | (3K4/(...)1, 6) [38] | 447,091.0 us | 2,146.22 us | 2,007.57 us |  1.41 |    0.01 | 17000.0000 |         - |   278497.8 KB |       0.260 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (3K4/(...)1, 6) [38] | 270,516.0 us | 1,031.18 us |   914.11 us |  0.85 |    0.01 | 17000.0000 |         - |  278497.45 KB |       0.260 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (3K4/(...)1, 6) [38] |  21,649.2 us |    58.64 us |    54.85 us |  0.07 |    0.00 |   312.5000 |         - |    5447.08 KB |       0.005 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (3K4/(...)1, 6) [38] |  21,053.5 us |   120.87 us |    94.36 us |  0.07 |    0.00 |   312.5000 |         - |    5447.08 KB |       0.005 |
  *  |                                                    |                      |              |             |             |       |         |            |           |               |             |
- *  | NewPosition                                        | (8/p7(...)-, 6) [37] |  12,793.4 us |   232.93 us |   217.88 us |  1.00 |    0.00 |  2625.0000 |   62.5000 |   42917.22 KB |        1.00 |
- *  | MakeUnmakeMove_Original                            | (8/p7(...)-, 6) [37] |  16,194.7 us |    88.10 us |    78.10 us |  1.27 |    0.02 |  1031.2500 |         - |   17115.45 KB |        0.40 |
- *  | MakeUnmakeMove_WithZobristKey                      | (8/p7(...)-, 6) [37] |  10,225.7 us |    72.37 us |    64.15 us |  0.80 |    0.01 |  1046.8750 |   15.6250 |   17115.44 KB |        0.40 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (8/p7(...)-, 6) [37] |   4,561.7 us |    30.06 us |    28.12 us |  0.36 |    0.01 |   507.8125 |    7.8125 |    8399.77 KB |        0.20 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (8/p7(...)-, 6) [37] |   4,648.8 us |    48.80 us |    45.65 us |  0.36 |    0.01 |   507.8125 |    7.8125 |    8399.77 KB |        0.20 |
+ *  | NewPosition                                        | (8/p7(...)-, 6) [37] |  13,088.8 us |   259.00 us |   242.27 us |  1.00 |    0.00 |  2625.0000 |   62.5000 |   42917.22 KB |        1.00 |
+ *  | MakeUnmakeMove_Original                            | (8/p7(...)-, 6) [37] |  16,455.7 us |   224.85 us |   210.32 us |  1.26 |    0.02 |  1031.2500 |         - |   17115.45 KB |        0.40 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (8/p7(...)-, 6) [37] |  10,077.0 us |    67.46 us |    63.10 us |  0.77 |    0.02 |  1046.8750 |   15.6250 |   17115.44 KB |        0.40 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (8/p7(...)-, 6) [37] |   4,534.2 us |    49.09 us |    45.92 us |  0.35 |    0.01 |   507.8125 |    7.8125 |    8399.77 KB |        0.20 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (8/p7(...)-, 6) [37] |   4,543.6 us |    46.58 us |    41.29 us |  0.35 |    0.01 |   507.8125 |    7.8125 |    8399.77 KB |        0.20 |
  *  |                                                    |                      |              |             |             |       |         |            |           |               |             |
- *  | NewPosition                                        | (r3k2(...)1, 4) [73] | 289,780.8 us | 3,074.26 us | 2,725.25 us | 1.000 |    0.00 | 49000.0000 |  500.0000 |  801232.84 KB |       1.000 |
- *  | MakeUnmakeMove_Original                            | (r3k2(...)1, 4) [73] | 355,328.7 us | 1,518.35 us | 1,420.27 us | 1.227 |    0.01 |  6000.0000 |         - |  104639.09 KB |       0.131 |
- *  | MakeUnmakeMove_WithZobristKey                      | (r3k2(...)1, 4) [73] | 242,989.5 us | 1,527.59 us | 1,428.91 us | 0.839 |    0.01 |  6333.3333 |         - |  104638.61 KB |       0.131 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (r3k2(...)1, 4) [73] |     160.4 us |     0.72 us |     0.67 us | 0.001 |    0.00 |     6.1035 |         - |     102.78 KB |       0.000 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (r3k2(...)1, 4) [73] |     168.7 us |     0.98 us |     0.87 us | 0.001 |    0.00 |     6.1035 |         - |     102.78 KB |       0.000 |
+ *  | NewPosition                                        | (r3k2(...)1, 4) [73] | 287,097.0 us | 5,167.07 us | 4,580.48 us | 1.000 |    0.00 | 49000.0000 |  500.0000 |  801232.84 KB |       1.000 |
+ *  | MakeUnmakeMove_Original                            | (r3k2(...)1, 4) [73] | 434,386.8 us | 1,566.22 us | 1,388.41 us | 1.513 |    0.02 |  6000.0000 |         - |  104639.09 KB |       0.131 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (r3k2(...)1, 4) [73] | 257,067.5 us |   987.63 us |   923.83 us | 0.895 |    0.01 |  6000.0000 |         - |  104638.73 KB |       0.131 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (r3k2(...)1, 4) [73] |     164.6 us |     0.64 us |     0.60 us | 0.001 |    0.00 |     6.1035 |         - |     102.78 KB |       0.000 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (r3k2(...)1, 4) [73] |     166.6 us |     2.61 us |     2.44 us | 0.001 |    0.00 |     6.1035 |         - |     102.78 KB |       0.000 |
  *  |                                                    |                      |              |             |             |       |         |            |           |               |             |
- *  | NewPosition                                        | (r4rk(...)0, 4) [77] | 257,510.9 us | 5,042.96 us | 4,717.19 us | 1.000 |    0.00 | 46500.0000 |  500.0000 |  765557.05 KB |       1.000 |
- *  | MakeUnmakeMove_Original                            | (r4rk(...)0, 4) [77] | 278,350.4 us |   753.56 us |   704.88 us | 1.081 |    0.02 |  5500.0000 |         - |   96332.82 KB |       0.126 |
- *  | MakeUnmakeMove_WithZobristKey                      | (r4rk(...)0, 4) [77] | 208,272.9 us |   985.54 us |   822.97 us | 0.810 |    0.01 |  5666.6667 |         - |    96332.7 KB |       0.126 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (r4rk(...)0, 4) [77] |     309.7 us |     2.04 us |     1.90 us | 0.001 |    0.00 |    12.6953 |         - |     208.53 KB |       0.000 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (r4rk(...)0, 4) [77] |     357.8 us |     3.60 us |     3.37 us | 0.001 |    0.00 |    12.6953 |         - |     208.53 KB |       0.000 |
+ *  | NewPosition                                        | (r4rk(...)0, 4) [77] | 265,197.6 us | 3,409.14 us | 3,188.91 us | 1.000 |    0.00 | 46500.0000 |  500.0000 |  765557.05 KB |       1.000 |
+ *  | MakeUnmakeMove_Original                            | (r4rk(...)0, 4) [77] | 275,395.6 us |   628.77 us |   557.39 us | 1.038 |    0.01 |  5500.0000 |         - |   96332.82 KB |       0.126 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (r4rk(...)0, 4) [77] | 206,780.7 us |   743.40 us |   580.39 us | 0.779 |    0.01 |  5666.6667 |         - |    96332.7 KB |       0.126 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (r4rk(...)0, 4) [77] |     347.7 us |     0.50 us |     0.39 us | 0.001 |    0.00 |    12.6953 |         - |     208.53 KB |       0.000 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (r4rk(...)0, 4) [77] |     341.5 us |     1.03 us |     0.97 us | 0.001 |    0.00 |    12.6953 |         - |     208.53 KB |       0.000 |
  *  |                                                    |                      |              |             |             |       |         |            |           |               |             |
- *  | NewPosition                                        | (rnbq(...)1, 4) [61] |  14,612.6 us |   221.92 us |   196.73 us |  1.00 |    0.00 |  2671.8750 |   31.2500 |   43734.68 KB |       1.000 |
- *  | MakeUnmakeMove_Original                            | (rnbq(...)1, 4) [61] |  15,658.0 us |    27.86 us |    24.70 us |  1.07 |    0.02 |   593.7500 |         - |    9762.01 KB |       0.223 |
- *  | MakeUnmakeMove_WithZobristKey                      | (rnbq(...)1, 4) [61] |  11,908.9 us |    38.41 us |    34.05 us |  0.82 |    0.01 |   593.7500 |         - |       9762 KB |       0.223 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (rnbq(...)1, 4) [61] |     225.9 us |     0.83 us |     0.70 us |  0.02 |    0.00 |    13.4277 |         - |     220.05 KB |       0.005 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (rnbq(...)1, 4) [61] |     256.0 us |     0.86 us |     0.81 us |  0.02 |    0.00 |    13.1836 |         - |     220.05 KB |       0.005 |
+ *  | NewPosition                                        | (rnbq(...)1, 4) [61] |  14,317.8 us |   137.40 us |   128.52 us |  1.00 |    0.00 |  2671.8750 |   31.2500 |   43734.68 KB |       1.000 |
+ *  | MakeUnmakeMove_Original                            | (rnbq(...)1, 4) [61] |  15,662.9 us |    37.47 us |    33.22 us |  1.09 |    0.01 |   593.7500 |         - |    9762.01 KB |       0.223 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (rnbq(...)1, 4) [61] |  11,864.3 us |    40.01 us |    35.47 us |  0.83 |    0.01 |   593.7500 |         - |       9762 KB |       0.223 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (rnbq(...)1, 4) [61] |     226.2 us |     1.48 us |     1.31 us |  0.02 |    0.00 |    13.4277 |         - |     220.05 KB |       0.005 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (rnbq(...)1, 4) [61] |     247.4 us |     1.40 us |     1.17 us |  0.02 |    0.00 |    13.1836 |         - |     220.05 KB |       0.005 |
  *
  *
  *  BenchmarkDotNet v0.13.12, macOS Monterey 12.7.2 (21G1974) [Darwin 21.6.0]
- *  Intel Xeon CPU E5-1650 v2 3.50GHz (Max: 3.34GHz), 1 CPU, 3 logical and 3 physical cores
+ *  Intel Core i7-8700B CPU 3.20GHz (Max: 3.19GHz) (Coffee Lake), 1 CPU, 4 logical and 4 physical cores
  *  .NET SDK 8.0.101
- *    [Host]     : .NET 8.0.1 (8.0.123.58001), X64 RyuJIT AVX
- *    DefaultJob : .NET 8.0.1 (8.0.123.58001), X64 RyuJIT AVX
+ *    [Host]     : .NET 8.0.1 (8.0.123.58001), X64 RyuJIT AVX2
+ *    DefaultJob : .NET 8.0.1 (8.0.123.58001), X64 RyuJIT AVX2
  *
  *  | Method                                             | data                 | Mean         | Error        | StdDev       | Median       | Ratio | RatioSD | Gen0        | Gen1      | Allocated     | Alloc Ratio |
  *  |--------------------------------------------------- |--------------------- |-------------:|-------------:|-------------:|-------------:|------:|--------:|------------:|----------:|--------------:|------------:|
- *  | NewPosition                                        | (2K2r(...)1, 6) [38] | 540,876.0 us | 10,642.51 us | 13,459.37 us | 538,884.9 us |  1.00 |    0.00 | 174000.0000 | 2000.0000 | 1069286.11 KB |        1.00 |
- *  | MakeUnmakeMove_Original                            | (2K2r(...)1, 6) [38] | 564,020.4 us |  4,597.97 us |  3,839.51 us | 562,780.9 us |  1.04 |    0.02 |  45000.0000 |         - |  278498.13 KB |        0.26 |
- *  | MakeUnmakeMove_WithZobristKey                      | (2K2r(...)1, 6) [38] | 466,090.7 us |  3,744.35 us |  3,502.46 us | 466,143.4 us |  0.86 |    0.02 |  45000.0000 |         - |  278498.13 KB |        0.26 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (2K2r(...)1, 6) [38] |  32,505.8 us |    618.37 us |    736.12 us |  32,372.7 us |  0.06 |    0.00 |   5687.5000 |   62.5000 |   35057.89 KB |        0.03 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (2K2r(...)1, 6) [38] |  33,390.2 us |    659.56 us |  1,317.21 us |  32,919.5 us |  0.06 |    0.00 |   5687.5000 |   62.5000 |   35057.89 KB |        0.03 |
+ *  | NewPosition                                        | (2K2r(...)1, 6) [38] | 596,556.4 us | 23,284.29 us | 67,921.36 us | 572,548.9 us |  1.00 |    0.00 | 174000.0000 | 2000.0000 | 1069286.11 KB |        1.00 |
+ *  | MakeUnmakeMove_Original                            | (2K2r(...)1, 6) [38] | 581,562.8 us | 32,111.07 us | 92,647.79 us | 569,710.1 us |  0.99 |    0.18 |  45000.0000 |         - |  278498.13 KB |        0.26 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (2K2r(...)1, 6) [38] | 385,904.5 us | 20,523.47 us | 60,513.91 us | 367,557.5 us |  0.66 |    0.12 |  45000.0000 |  500.0000 |  278497.61 KB |        0.26 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (2K2r(...)1, 6) [38] |  26,102.9 us |    553.67 us |  1,552.56 us |  25,638.0 us |  0.04 |    0.01 |   5718.7500 |   62.5000 |   35057.85 KB |        0.03 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (2K2r(...)1, 6) [38] |  23,625.7 us |    431.98 us |    697.56 us |  23,592.0 us |  0.04 |    0.00 |   5718.7500 |   62.5000 |   35057.85 KB |        0.03 |
  *  |                                                    |                      |              |              |              |              |       |         |             |           |               |             |
- *  | NewPosition                                        | (3K4/(...)1, 6) [38] | 548,045.1 us | 10,821.74 us | 15,862.37 us | 549,152.9 us |  1.00 |    0.00 | 174000.0000 | 2000.0000 | 1069286.11 KB |       1.000 |
- *  | MakeUnmakeMove_Original                            | (3K4/(...)1, 6) [38] | 556,688.0 us |  5,922.58 us |  4,945.62 us | 555,281.0 us |  1.02 |    0.02 |  45000.0000 |         - |  278498.13 KB |       0.260 |
- *  | MakeUnmakeMove_WithZobristKey                      | (3K4/(...)1, 6) [38] | 463,567.3 us |  2,885.32 us |  2,698.93 us | 463,340.7 us |  0.86 |    0.02 |  45000.0000 |         - |  278498.13 KB |       0.260 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (3K4/(...)1, 6) [38] |  31,099.1 us |    603.40 us |    991.40 us |  30,695.8 us |  0.06 |    0.00 |    875.0000 |         - |    5448.71 KB |       0.005 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (3K4/(...)1, 6) [38] |  30,692.5 us |    575.28 us |  1,411.18 us |  30,427.9 us |  0.06 |    0.00 |    833.3333 |         - |    5448.63 KB |       0.005 |
+ *  | NewPosition                                        | (3K4/(...)1, 6) [38] | 520,639.0 us | 11,233.39 us | 31,499.62 us | 515,944.3 us |  1.00 |    0.00 | 174000.0000 | 2000.0000 | 1069286.11 KB |       1.000 |
+ *  | MakeUnmakeMove_Original                            | (3K4/(...)1, 6) [38] | 487,835.4 us |  9,495.67 us | 21,433.32 us | 481,627.1 us |  0.94 |    0.07 |  45000.0000 |         - |  278498.13 KB |       0.260 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (3K4/(...)1, 6) [38] | 369,615.4 us | 13,953.87 us | 39,584.82 us | 365,639.3 us |  0.71 |    0.08 |  45000.0000 |  500.0000 |  278497.61 KB |       0.260 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (3K4/(...)1, 6) [38] |  27,328.3 us |  1,008.16 us |  2,908.77 us |  26,738.7 us |  0.05 |    0.01 |    875.0000 |         - |    5448.71 KB |       0.005 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (3K4/(...)1, 6) [38] |  23,015.6 us |    553.61 us |  1,606.12 us |  22,453.1 us |  0.04 |    0.00 |    875.0000 |         - |    5448.71 KB |       0.005 |
  *  |                                                    |                      |              |              |              |              |       |         |             |           |               |             |
- *  | NewPosition                                        | (8/p7(...)-, 6) [37] |  20,937.3 us |    387.25 us |    362.23 us |  20,862.1 us |  1.00 |    0.00 |   7000.0000 |  125.0000 |   42917.24 KB |        1.00 |
- *  | MakeUnmakeMove_Original                            | (8/p7(...)-, 6) [37] |  21,898.1 us |    419.94 us |    372.27 us |  21,922.7 us |  1.05 |    0.03 |   2781.2500 |   31.2500 |   17115.46 KB |        0.40 |
- *  | MakeUnmakeMove_WithZobristKey                      | (8/p7(...)-, 6) [37] |  18,003.7 us |    334.82 us |    313.19 us |  18,006.5 us |  0.86 |    0.02 |   2781.2500 |   31.2500 |   17115.46 KB |        0.40 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (8/p7(...)-, 6) [37] |   6,907.5 us |    137.10 us |    192.20 us |   6,890.6 us |  0.33 |    0.01 |   1367.1875 |   23.4375 |    8401.25 KB |        0.20 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (8/p7(...)-, 6) [37] |   9,188.5 us |    495.81 us |  1,461.91 us |   9,247.5 us |  0.41 |    0.07 |   1367.1875 |   23.4375 |    8401.25 KB |        0.20 |
+ *  | NewPosition                                        | (8/p7(...)-, 6) [37] |  18,770.8 us |    371.77 us |    890.73 us |  18,935.4 us |  1.00 |    0.00 |   7000.0000 |  125.0000 |   42917.24 KB |        1.00 |
+ *  | MakeUnmakeMove_Original                            | (8/p7(...)-, 6) [37] |  17,203.1 us |    340.58 us |    782.54 us |  17,155.4 us |  0.92 |    0.06 |   2781.2500 |   31.2500 |   17115.46 KB |        0.40 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (8/p7(...)-, 6) [37] |  12,017.8 us |    239.95 us |    344.13 us |  12,025.8 us |  0.65 |    0.03 |   2781.2500 |   46.8750 |   17115.45 KB |        0.40 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (8/p7(...)-, 6) [37] |   5,453.7 us |    106.75 us |    109.62 us |   5,443.5 us |  0.30 |    0.02 |   1367.1875 |   23.4375 |    8401.25 KB |        0.20 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (8/p7(...)-, 6) [37] |   5,295.0 us |    105.24 us |    195.07 us |   5,260.0 us |  0.29 |    0.02 |   1367.1875 |   23.4375 |    8401.25 KB |        0.20 |
  *  |                                                    |                      |              |              |              |              |       |         |             |           |               |             |
- *  | NewPosition                                        | (r3k2(...)1, 4) [73] | 572,501.0 us | 28,498.69 us | 84,029.03 us | 573,923.6 us | 1.000 |    0.00 | 130000.0000 | 1000.0000 |  801233.53 KB |       1.000 |
- *  | MakeUnmakeMove_Original                            | (r3k2(...)1, 4) [73] | 663,197.3 us | 33,037.67 us | 97,412.33 us | 645,325.6 us | 1.188 |    0.27 |  17000.0000 |         - |  104639.41 KB |       0.131 |
- *  | MakeUnmakeMove_WithZobristKey                      | (r3k2(...)1, 4) [73] | 419,634.7 us | 23,843.13 us | 68,792.89 us | 393,369.1 us | 0.749 |    0.17 |  17000.0000 |         - |  104639.41 KB |       0.131 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (r3k2(...)1, 4) [73] |     267.1 us |     13.87 us |     40.25 us |     256.7 us | 0.000 |    0.00 |     16.6016 |         - |     102.81 KB |       0.000 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (r3k2(...)1, 4) [73] |     232.9 us |      4.65 us |     10.58 us |     231.6 us | 0.000 |    0.00 |     16.6016 |         - |     102.81 KB |       0.000 |
+ *  | NewPosition                                        | (r3k2(...)1, 4) [73] | 521,874.5 us | 27,099.15 us | 78,187.23 us | 481,362.9 us | 1.000 |    0.00 | 130000.0000 | 1000.0000 |  801233.53 KB |       1.000 |
+ *  | MakeUnmakeMove_Original                            | (r3k2(...)1, 4) [73] | 367,470.0 us |  4,942.33 us |  6,426.42 us | 366,073.7 us | 0.686 |    0.10 |  17000.0000 |         - |  104639.41 KB |       0.131 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (r3k2(...)1, 4) [73] | 301,125.9 us |  3,070.02 us |  2,396.87 us | 300,703.8 us | 0.594 |    0.06 |  17000.0000 |         - |  104638.89 KB |       0.131 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (r3k2(...)1, 4) [73] |     163.3 us |      2.86 us |      2.39 us |     164.2 us | 0.000 |    0.00 |     16.6016 |         - |     102.81 KB |       0.000 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (r3k2(...)1, 4) [73] |     168.6 us |      2.50 us |      2.22 us |     168.4 us | 0.000 |    0.00 |     16.6016 |         - |     102.81 KB |       0.000 |
  *  |                                                    |                      |              |              |              |              |       |         |             |           |               |             |
- *  | NewPosition                                        | (r4rk(...)0, 4) [77] | 485,734.5 us |  9,694.63 us | 25,708.79 us | 482,125.0 us | 1.000 |    0.00 | 124000.0000 | 1000.0000 |  765557.73 KB |       1.000 |
- *  | MakeUnmakeMove_Original                            | (r4rk(...)0, 4) [77] | 511,174.7 us | 10,162.18 us | 24,347.94 us | 515,608.8 us | 1.059 |    0.07 |  15000.0000 |         - |   96333.51 KB |       0.126 |
- *  | MakeUnmakeMove_WithZobristKey                      | (r4rk(...)0, 4) [77] | 324,566.1 us |  6,379.08 us | 11,981.46 us | 321,145.3 us | 0.675 |    0.05 |  15500.0000 |         - |   96332.98 KB |       0.126 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (r4rk(...)0, 4) [77] |     535.0 us |     10.60 us |      9.92 us |     535.7 us | 0.001 |    0.00 |     33.2031 |         - |      208.6 KB |       0.000 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (r4rk(...)0, 4) [77] |     531.5 us |      8.47 us |      7.92 us |     530.2 us | 0.001 |    0.00 |     33.2031 |         - |      208.6 KB |       0.000 |
+ *  | NewPosition                                        | (r4rk(...)0, 4) [77] | 386,108.7 us |  7,594.60 us |  8,126.14 us | 387,721.3 us | 1.000 |    0.00 | 124000.0000 | 1000.0000 |  765557.73 KB |       1.000 |
+ *  | MakeUnmakeMove_Original                            | (r4rk(...)0, 4) [77] | 308,136.0 us |  5,898.38 us |  6,792.58 us | 307,536.4 us | 0.798 |    0.03 |  15500.0000 |         - |   96332.98 KB |       0.126 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (r4rk(...)0, 4) [77] | 236,499.4 us |  3,099.38 us |  2,899.16 us | 237,067.0 us | 0.610 |    0.01 |  15666.6667 |         - |   96332.81 KB |       0.126 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (r4rk(...)0, 4) [77] |     312.4 us |      6.07 us |      5.67 us |     312.1 us | 0.001 |    0.00 |     33.6914 |         - |      208.6 KB |       0.000 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (r4rk(...)0, 4) [77] |     363.2 us |      6.15 us |      5.75 us |     362.3 us | 0.001 |    0.00 |     33.6914 |         - |      208.6 KB |       0.000 |
  *  |                                                    |                      |              |              |              |              |       |         |             |           |               |             |
- *  | NewPosition                                        | (rnbq(...)1, 4) [61] |  25,527.8 us |    353.02 us |    312.94 us |  25,563.4 us |  1.00 |    0.00 |   7125.0000 |   93.7500 |    43734.7 KB |       1.000 |
- *  | MakeUnmakeMove_Original                            | (rnbq(...)1, 4) [61] |  26,917.2 us |    524.57 us |    515.20 us |  26,961.6 us |  1.05 |    0.03 |   1593.7500 |         - |    9763.26 KB |       0.223 |
- *  | MakeUnmakeMove_WithZobristKey                      | (rnbq(...)1, 4) [61] |  17,795.6 us |    305.66 us |    386.56 us |  17,654.9 us |  0.70 |    0.01 |   1593.7500 |         - |    9763.26 KB |       0.223 |
- *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (rnbq(...)1, 4) [61] |     363.8 us |      5.50 us |      5.14 us |     363.5 us |  0.01 |    0.00 |     35.6445 |         - |     220.12 KB |       0.005 |
- *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (rnbq(...)1, 4) [61] |     371.4 us |      7.35 us |      8.17 us |     370.2 us |  0.01 |    0.00 |     35.6445 |         - |     220.12 KB |       0.005 |
+ *  | NewPosition                                        | (rnbq(...)1, 4) [61] |  20,415.4 us |    380.05 us |    355.50 us |  20,395.0 us |  1.00 |    0.00 |   7125.0000 |   93.7500 |    43734.7 KB |       1.000 |
+ *  | MakeUnmakeMove_Original                            | (rnbq(...)1, 4) [61] |  16,943.5 us |    144.12 us |    120.35 us |  16,999.3 us |  0.83 |    0.02 |   1593.7500 |         - |    9763.26 KB |       0.223 |
+ *  | MakeUnmakeMove_WithZobristKey                      | (rnbq(...)1, 4) [61] |  13,816.4 us |    273.34 us |    355.42 us |  13,743.5 us |  0.69 |    0.02 |   1593.7500 |   15.6250 |    9763.24 KB |       0.223 |
+ *  | MakeUnmakeMove_WithZobristKey_PreSwitchSpecialMove | (rnbq(...)1, 4) [61] |     251.5 us |      3.29 us |      2.91 us |     251.1 us |  0.01 |    0.00 |     35.6445 |         - |     220.12 KB |       0.005 |
+ *  | MakeUnmakeMove_WithZobristKey_SwitchSpecialMove    | (rnbq(...)1, 4) [61] |     316.3 us |      2.19 us |      2.05 us |     315.9 us |  0.02 |    0.00 |     35.6445 |         - |     220.12 KB |       0.005 |
+ *
  */
 
 #pragma warning disable S101, S1854 // Types should be named in PascalCase
